+++ src/main/java/roidrole/patternbanners/core/PatternBannersCore.java
@@ -44,7 +44,7 @@ public class PatternBannersCore implements IFMLLoadingPlugin, IEarlyMixinLoader 
     @Override
     public List<String> getMixinConfigs() {
         List<String> mixinconfs = new ArrayList<>(2);
-        mixinconfs.add("mixins.patternbanners.json");
+        if(ConfigCore.wearBannerOnHead){mixinconfs.add("mixins.patternbanners.wear_banners.json");}
         if(ConfigCore.no_vanilla_pattern_apply){mixinconfs.add("mixins.patternbanners.no_vanilla_pattern_apply.json");}
         return mixinconfs;
     }